KeyProcessing lock implemented for single and bundled requests
