+++ osgp/protocol-adapter-dlms/osgp-protocol-adapter-dlms/src/main/java/org/opensmartgridplatform/adapter/protocol/dlms/domain/commands/security/ReplaceKeyCommandExecutor.java
@@ -8,26 +8,12 @@
  */
 package org.opensmartgridplatform.adapter.protocol.dlms.domain.commands.security;
 
-import java.io.IOException;
-import org.openmuc.jdlms.MethodParameter;
-import org.openmuc.jdlms.MethodResultCode;
-import org.openmuc.jdlms.SecurityUtils;
-import org.openmuc.jdlms.SecurityUtils.KeyId;
-import org.opensmartgridplatform.adapter.protocol.dlms.application.services.SecretManagementService;
-import org.opensmartgridplatform.adapter.protocol.dlms.domain.commands.AbstractCommandExecutor;
-import org.opensmartgridplatform.adapter.protocol.dlms.domain.commands.utils.JdlmsObjectToStringUtil;
+import org.opensmartgridplatform.adapter.protocol.dlms.application.services.DeviceKeyProcessingService;
 import org.opensmartgridplatform.adapter.protocol.dlms.domain.entities.DlmsDevice;
-import org.opensmartgridplatform.adapter.protocol.dlms.domain.entities.SecurityKeyType;
 import org.opensmartgridplatform.adapter.protocol.dlms.domain.factories.DlmsConnectionManager;
-import org.opensmartgridplatform.adapter.protocol.dlms.exceptions.ConnectionException;
-import org.opensmartgridplatform.adapter.protocol.dlms.exceptions.ProtocolAdapterException;
-import org.opensmartgridplatform.dto.valueobjects.smartmetering.ActionRequestDto;
+import org.opensmartgridplatform.adapter.protocol.dlms.exceptions.DeviceKeyProcessAlreadyRunningException;
 import org.opensmartgridplatform.dto.valueobjects.smartmetering.ActionResponseDto;
 import org.opensmartgridplatform.dto.valueobjects.smartmetering.SetKeysRequestDto;
-import org.opensmartgridplatform.shared.exceptionhandling.ComponentType;
-import org.opensmartgridplatform.shared.exceptionhandling.EncrypterException;
-import org.opensmartgridplatform.shared.exceptionhandling.FunctionalException;
-import org.opensmartgridplatform.shared.exceptionhandling.FunctionalExceptionType;
 import org.opensmartgridplatform.shared.exceptionhandling.OsgpException;
 import org.opensmartgridplatform.shared.infra.jms.MessageMetadata;
 import org.opensmartgridplatform.shared.security.RsaEncrypter;
@@ -49,14 +35,11 @@ import org.springframework.stereotype.Component;
  */
 @Component
 public class ReplaceKeyCommandExecutor
-    extends AbstractCommandExecutor<ReplaceKeyInput, DlmsDevice> {
+    extends AbstractReplaceKeyCommandExecutor<SetKeysRequestDto> {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(ReplaceKeyCommandExecutor.class);
 
-  private static final String REPLACE_KEYS = "Replace keys for device: ";
-  private static final String WAS_SUCCESFULL = " was successful";
-
-  @Autowired private SecretManagementService secretManagementService;
+  @Autowired private DeviceKeyProcessingService deviceKeyProcessingService;
 
   @Autowired
   @Qualifier("decrypterForGxfSmartMetering")
@@ -66,146 +49,29 @@ public class ReplaceKeyCommandExecutor
     super(SetKeysRequestDto.class);
   }
 
-  public static ReplaceKeyInput wrap(
-      final byte[] bytes,
-      final KeyId keyId,
-      final SecurityKeyType securityKeyType,
-      final boolean isGenerated) {
-    return new ReplaceKeyInput(bytes, keyId, securityKeyType, isGenerated);
-  }
-
   @Override
-  public ActionResponseDto executeBundleAction(
+  public ActionResponseDto execute(
       final DlmsConnectionManager conn,
       final DlmsDevice device,
-      final ActionRequestDto actionRequestDto,
+      final SetKeysRequestDto setKeysRequestDto,
       final MessageMetadata messageMetadata)
       throws OsgpException {
-
-    this.checkActionRequestType(actionRequestDto);
-
-    LOGGER.info("Keys set on device :{}", device.getDeviceIdentification());
-
-    SetKeysRequestDto setKeysRequestDto = (SetKeysRequestDto) actionRequestDto;
-
-    if (!setKeysRequestDto.isGeneratedKeys()) {
-      setKeysRequestDto = this.decryptRsaKeys((SetKeysRequestDto) actionRequestDto);
-    }
-    // if keys are generated, then they are unencrypted by the GenerateAndReplaceKeyCommandExecutor
-
-    final DlmsDevice devicePostSave =
-        this.execute(
-            conn,
-            device,
-            ReplaceKeyCommandExecutor.wrap(
-                setKeysRequestDto.getAuthenticationKey(),
-                KeyId.AUTHENTICATION_KEY,
-                SecurityKeyType.E_METER_AUTHENTICATION,
-                setKeysRequestDto.isGeneratedKeys()),
-            messageMetadata);
-
-    this.execute(
-        conn,
-        devicePostSave,
-        ReplaceKeyCommandExecutor.wrap(
-            setKeysRequestDto.getEncryptionKey(),
-            KeyId.GLOBAL_UNICAST_ENCRYPTION_KEY,
-            SecurityKeyType.E_METER_ENCRYPTION,
-            setKeysRequestDto.isGeneratedKeys()),
-        messageMetadata);
-
-    return new ActionResponseDto(REPLACE_KEYS + device.getDeviceIdentification() + WAS_SUCCESFULL);
-  }
-
-  private SetKeysRequestDto decryptRsaKeys(final SetKeysRequestDto setKeysRequestDto)
-      throws FunctionalException {
-    try {
-      final byte[] authenticationKey =
-          this.decrypterForGxfSmartMetering.decrypt(setKeysRequestDto.getAuthenticationKey());
-      final byte[] encryptionKey =
-          this.decrypterForGxfSmartMetering.decrypt(setKeysRequestDto.getEncryptionKey());
-
-      return new SetKeysRequestDto(authenticationKey, encryptionKey);
-    } catch (final EncrypterException e) {
-      throw new FunctionalException(
-          FunctionalExceptionType.DECRYPTION_EXCEPTION, ComponentType.PROTOCOL_DLMS, e);
-    }
-  }
-
-  @Override
-  public DlmsDevice execute(
-      final DlmsConnectionManager conn,
-      final DlmsDevice device,
-      final ReplaceKeyInput keyWrapper,
-      final MessageMetadata messageMetadata)
-      throws OsgpException {
-
-    if (!keyWrapper.isGenerated()) {
-      this.secretManagementService.storeNewKey(
-          messageMetadata,
-          device.getDeviceIdentification(),
-          keyWrapper.getSecurityKeyType(),
-          keyWrapper.getBytes());
-    }
-
-    this.sendToDevice(conn, device.getDeviceIdentification(), keyWrapper, messageMetadata);
-    this.secretManagementService.activateNewKey(
-        messageMetadata, device.getDeviceIdentification(), keyWrapper.getSecurityKeyType());
-    return device;
-  }
-
-  /**
-   * Send the key to the device.
-   *
-   * @param conn jDLMS connection.
-   * @param deviceIdentification Device identification
-   * @param keyWrapper Key data
-   * @param messageMetadata
-   */
-  private void sendToDevice(
-      final DlmsConnectionManager conn,
-      final String deviceIdentification,
-      final ReplaceKeyInput keyWrapper,
-      final MessageMetadata messageMetadata)
-      throws ProtocolAdapterException {
+    this.checkActionRequestType(setKeysRequestDto);
 
     try {
-      final byte[] decryptedKey = keyWrapper.getBytes();
-      final byte[] decryptedMasterKey =
-          this.secretManagementService.getKey(
-              messageMetadata, deviceIdentification, SecurityKeyType.E_METER_MASTER);
+      this.deviceKeyProcessingService.startProcessing(messageMetadata.getDeviceIdentification());
+    } catch (final DeviceKeyProcessAlreadyRunningException e) {
+      // This exception will be caught in the DeviceRequestMessageProcessor.
+      // The request will NOT be sent back to Core to retry but put back on the queue
+      LOGGER.info(
+          "Key changing process already running on device :{}", device.getDeviceIdentification());
+      throw e;
+    }
 
-      final MethodParameter methodParameterAuth =
-          SecurityUtils.keyChangeMethodParamFor(
-              decryptedMasterKey, decryptedKey, keyWrapper.getKeyId());
-
-      conn.getDlmsMessageListener()
-          .setDescription(
-              "ReplaceKey for "
-                  + keyWrapper.getSecurityKeyType()
-                  + " "
-                  + keyWrapper.getKeyId()
-                  + ", call method: "
-                  + JdlmsObjectToStringUtil.describeMethod(methodParameterAuth));
-
-      final MethodResultCode methodResultCode =
-          conn.getConnection().action(methodParameterAuth).getResultCode();
-
-      if (!MethodResultCode.SUCCESS.equals(methodResultCode)) {
-        throw new ProtocolAdapterException(
-            "AccessResultCode for replace keys was not SUCCESS: " + methodResultCode);
-      }
-
-      if (keyWrapper.getSecurityKeyType() == SecurityKeyType.E_METER_AUTHENTICATION) {
-        conn.getConnection().changeClientGlobalAuthenticationKey(decryptedKey);
-      } else if (keyWrapper.getSecurityKeyType() == SecurityKeyType.E_METER_ENCRYPTION) {
-        conn.getConnection().changeClientGlobalEncryptionKey(decryptedKey);
-      }
-    } catch (final IOException e) {
-      throw new ConnectionException(e);
-    } catch (final EncrypterException e) {
-      throw new ProtocolAdapterException(
-          "Unexpected exception during decryption of security keys, reason = " + e.getMessage(), e);
+    try {
+      return this.replaceKeys(conn, device, setKeysRequestDto, messageMetadata);
+    } finally {
+      this.deviceKeyProcessingService.stopProcessing(messageMetadata.getDeviceIdentification());
     }
   }
 }
